requirements.txt
py==1.11.0
py-clob-client==0.14.0
python-dotenv==0.19.2
requests==2.27.1
